Separates websocket stuff from event stuff.
Passes the websocket send function in partially applied with the necessary
connection object to give the events code a function that only takes a message
to be sent to the play, which is all the events code should care about. This
paves the way for altnerate connection types--telnet, for example!

Signed-off-by: Wayne <wayne@puppetlabs.com>