Adds basic telnet support.
This isn't really telnet, but line-delimited TCP which corresponds to a simple
"line-by-line" fallback mode in the telnet protocol. I think this may need to be
modified to allow simple telnet clients to negotiate non-interruptive
asynchronous messages from the server--for example, messages sent by
events/notify-players. I think the primary use case for this telnet interface
though is to facilitate testing.

Signed-off-by: Wayne <wayne@puppetlabs.com>